+++ autodocodec-api-usage/test_resources/nix/local-time-type.nix
@@ -1,1 +1,1 @@
-types.unspecified
+types.str