+++ autodocodec-api-usage/test_resources/nix/monoid-first-type.nix
@@ -1,0 +1,4 @@
+types.oneOf  [
+  types.anything
+  types.str
+]