+++ autodocodec-nix/CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
-## [0.0.0.0] 
+## [0.0.0.0] - 2024-07-19
+
+Initial version